Prepared vision for merge, added command references
